```
       django_template.py             6.4s (2)             6.3s (2)  -1.4%
            pyxl_bench.py             4.6s (2)             4.4s (2)  -4.4%
 sqlalchemy_imperative.py             2.4s (2)             2.4s (2)  +0.3%
        django_migrate.py             2.0s (2)             2.0s (2)  -1.2%
      virtualenv_bench.py             8.2s (2)             8.2s (2)  -0.8%
                  geomean                 4.1s                 4.0s  -1.5%
```